rename chapter6 to chapter4
