+++ src/main/webapp/chapter4/todolist-mvc.zul
@@ -1,5 +1,5 @@
 <?link rel="stylesheet" type="text/css" href="/style.css"?>
-<window apply="org.zkoss.essentials.chapter6.mvc.TodoListController" 
+<window apply="org.zkoss.essentials.chapter4.mvc.TodoListController" 
 	border="normal" hflex="1" vflex="1" contentStyle="overflow:auto">
 	<caption src="/imgs/todo.png" sclass="fn-caption" label="Todo List (MVC)"/>
 	<borderlayout>
