+++ src/main/webapp/chapter7/pagebased/index-todolist-mvc.zul
@@ -3,6 +3,6 @@
 <div height="100%">
 	<template name="layout" src="/chapter7/pagebased/layout/template.zul"/>
 	<apply template="layout"/>
-	<template name="center" src="/chapter6/todolist-mvc.zul"/>
+	<template name="center" src="/chapter4/todolist-mvc.zul"/>
 </div>
 </zk>
