+++ src/main/webapp/index.zul
@@ -26,8 +26,8 @@
 				
 				<label value="Chapter 6" sclass="head2"/>
 				<hlayout>
-					<a href="/chapter6/index.zul">MVC</a>
-					<a href="/chapter6/index-mvvm.zul">MVVM</a>
+					<a href="/chapter4/index.zul">MVC</a>
+					<a href="/chapter4/index-mvvm.zul">MVVM</a>
 				</hlayout>
 				
 				<label value="Chapter 7" sclass="head2"/>
